+++ modules/util/portal-tools-sample-sql-builder/src/main/java/com/liferay/portal/tools/sample/sql/builder/DDLDDMDataFactory.java
@@ -1,0 +1,293 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.tools.sample.sql.builder;
+
+import com.liferay.document.library.kernel.model.DLFileEntryMetadata;
+import com.liferay.document.library.kernel.model.DLFileEntryMetadataModel;
+import com.liferay.dynamic.data.lists.model.DDLRecordConstants;
+import com.liferay.dynamic.data.lists.model.DDLRecordModel;
+import com.liferay.dynamic.data.lists.model.DDLRecordSet;
+import com.liferay.dynamic.data.lists.model.DDLRecordSetConstants;
+import com.liferay.dynamic.data.lists.model.DDLRecordSetModel;
+import com.liferay.dynamic.data.lists.model.DDLRecordVersionModel;
+import com.liferay.dynamic.data.lists.model.impl.DDLRecordModelImpl;
+import com.liferay.dynamic.data.lists.model.impl.DDLRecordSetModelImpl;
+import com.liferay.dynamic.data.lists.model.impl.DDLRecordVersionModelImpl;
+import com.liferay.dynamic.data.mapping.model.DDMContentModel;
+import com.liferay.dynamic.data.mapping.model.DDMStructureLayoutModel;
+import com.liferay.dynamic.data.mapping.model.DDMStructureLinkModel;
+import com.liferay.dynamic.data.mapping.model.DDMStructureModel;
+import com.liferay.dynamic.data.mapping.model.DDMStructureVersionModel;
+import com.liferay.dynamic.data.mapping.model.impl.DDMStructureLinkModelImpl;
+import com.liferay.petra.string.StringPool;
+import com.liferay.portal.kernel.util.StringBundler;
+import com.liferay.portal.kernel.util.StringUtil;
+import com.liferay.portal.kernel.workflow.WorkflowConstants;
+
+import java.util.Date;
+
+/**
+ * @author Lily Chi
+ */
+public class DDLDDMDataFactory extends BaseDDMDataFactory {
+
+	public DDLDDMDataFactory() {
+		_defaultDDLDDMStructureVersionId = counter.get();
+	}
+
+	public int getMaxDDLRecordCount() {
+		return BenchmarksPropsValues.MAX_DDL_RECORD_COUNT;
+	}
+
+	public int getMaxDDLRecordSetCount() {
+		return BenchmarksPropsValues.MAX_DDL_RECORD_SET_COUNT;
+	}
+
+	public DDMStructureLayoutModel newDDLDDMStructureLayoutModel(
+		long groupId, DDMStructureVersionModel ddmStructureVersionModel) {
+
+		StringBundler sb = new StringBundler(
+			3 + BenchmarksPropsValues.MAX_DDL_CUSTOM_FIELD_COUNT * 4);
+
+		sb.append("{\"defaultLanguageId\": \"en_US\", \"pages\": [{\"rows\": ");
+		sb.append("[");
+
+		for (int i = 0; i < BenchmarksPropsValues.MAX_DDL_CUSTOM_FIELD_COUNT;
+			 i++) {
+
+			sb.append("{\"columns\": [{\"fieldNames\": [\"");
+			sb.append(nextDDLCustomFieldName(groupId, i));
+			sb.append("\"], \"size\": 12}]}");
+			sb.append(", ");
+		}
+
+		if (BenchmarksPropsValues.MAX_DDL_CUSTOM_FIELD_COUNT > 0) {
+			sb.setIndex(sb.index() - 1);
+		}
+
+		sb.append("], \"title\": {\"en_US\": \"\"}}],\"paginationMode\": ");
+		sb.append("\"single-page\"}");
+
+		return newDDMStructureLayoutModel(
+			GLOBAL_GROUP_ID, DEFAULT_USER_ID,
+			ddmStructureVersionModel.getStructureVersionId(), sb.toString());
+	}
+
+	public DDMStructureModel newDDLDDMStructureModel(long groupId) {
+		StringBundler sb = new StringBundler(
+			3 + BenchmarksPropsValues.MAX_DDL_CUSTOM_FIELD_COUNT * 9);
+
+		sb.append("{\"availableLanguageIds\": [\"en_US\"],");
+		sb.append("\"defaultLanguageId\": \"en_US\", \"fields\": [");
+
+		for (int i = 0; i < BenchmarksPropsValues.MAX_DDL_CUSTOM_FIELD_COUNT;
+			 i++) {
+
+			sb.append("{\"dataType\": \"string\", \"indexType\": ");
+			sb.append("\"keyword\", \"label\": {\"en_US\": \"Text");
+			sb.append(i);
+			sb.append("\"}, \"name\": \"");
+			sb.append(nextDDLCustomFieldName(groupId, i));
+			sb.append("\", \"readOnly\": false, \"repeatable\": false,");
+			sb.append("\"required\": false, \"showLabel\": true, \"type\": ");
+			sb.append("\"text\"}");
+			sb.append(",");
+		}
+
+		if (BenchmarksPropsValues.MAX_DDL_CUSTOM_FIELD_COUNT > 0) {
+			sb.setIndex(sb.index() - 1);
+		}
+
+		sb.append("]}");
+
+		return newDDMStructureModel(
+			groupId, SAMPLE_USER_ID, getClassNameId(DDLRecordSet.class),
+			DataFactoryConstants.DDL_DDM_STRUCTURE_KEY, sb.toString(),
+			counter.get());
+	}
+
+	public DDLRecordModel newDDLRecordModel(
+		DDLRecordSetModel dDLRecordSetModel) {
+
+		DDLRecordModel ddlRecordModel = new DDLRecordModelImpl();
+
+		ddlRecordModel.setUuid(SequentialUUID.generate());
+		ddlRecordModel.setRecordId(counter.get());
+		ddlRecordModel.setGroupId(dDLRecordSetModel.getGroupId());
+		ddlRecordModel.setCompanyId(COMPANY_ID);
+		ddlRecordModel.setUserId(SAMPLE_USER_ID);
+		ddlRecordModel.setUserName(DataFactoryConstants.SAMPLE_USER_NAME);
+		ddlRecordModel.setVersionUserId(SAMPLE_USER_ID);
+		ddlRecordModel.setVersionUserName(
+			DataFactoryConstants.SAMPLE_USER_NAME);
+		ddlRecordModel.setCreateDate(new Date());
+		ddlRecordModel.setModifiedDate(new Date());
+		ddlRecordModel.setDDMStorageId(counter.get());
+		ddlRecordModel.setRecordSetId(dDLRecordSetModel.getRecordSetId());
+		ddlRecordModel.setVersion(DDLRecordConstants.VERSION_DEFAULT);
+		ddlRecordModel.setDisplayIndex(
+			DDLRecordConstants.DISPLAY_INDEX_DEFAULT);
+		ddlRecordModel.setLastPublishDate(new Date());
+
+		return ddlRecordModel;
+	}
+
+	public DDLRecordSetModel newDDLRecordSetModel(
+		DDMStructureModel ddmStructureModel, int currentIndex) {
+
+		DDLRecordSetModel ddlRecordSetModel = new DDLRecordSetModelImpl();
+
+		ddlRecordSetModel.setUuid(SequentialUUID.generate());
+		ddlRecordSetModel.setRecordSetId(counter.get());
+		ddlRecordSetModel.setGroupId(ddmStructureModel.getGroupId());
+		ddlRecordSetModel.setCompanyId(COMPANY_ID);
+		ddlRecordSetModel.setUserId(SAMPLE_USER_ID);
+		ddlRecordSetModel.setUserName(DataFactoryConstants.SAMPLE_USER_NAME);
+		ddlRecordSetModel.setCreateDate(new Date());
+		ddlRecordSetModel.setModifiedDate(new Date());
+		ddlRecordSetModel.setDDMStructureId(ddmStructureModel.getStructureId());
+		ddlRecordSetModel.setRecordSetKey(String.valueOf(counter.get()));
+
+		StringBundler sb = new StringBundler(5);
+
+		sb.append("<?xml version=\"1.0\"?><root available-locales=\"en_US\" ");
+		sb.append("default-locale=\"en_US\"><name language-id=\"en_US\">");
+		sb.append("Test DDL Record Set ");
+		sb.append(currentIndex);
+		sb.append("</name></root>");
+
+		ddlRecordSetModel.setName(sb.toString());
+
+		ddlRecordSetModel.setMinDisplayRows(
+			DDLRecordSetConstants.MIN_DISPLAY_ROWS_DEFAULT);
+		ddlRecordSetModel.setScope(
+			DDLRecordSetConstants.SCOPE_DYNAMIC_DATA_LISTS);
+		ddlRecordSetModel.setSettings(StringPool.BLANK);
+		ddlRecordSetModel.setLastPublishDate(new Date());
+
+		return ddlRecordSetModel;
+	}
+
+	public DDLRecordVersionModel newDDLRecordVersionModel(
+		DDLRecordModel dDLRecordModel) {
+
+		DDLRecordVersionModel ddlRecordVersionModel =
+			new DDLRecordVersionModelImpl();
+
+		ddlRecordVersionModel.setRecordVersionId(counter.get());
+		ddlRecordVersionModel.setGroupId(dDLRecordModel.getGroupId());
+		ddlRecordVersionModel.setCompanyId(COMPANY_ID);
+		ddlRecordVersionModel.setUserId(SAMPLE_USER_ID);
+		ddlRecordVersionModel.setUserName(
+			DataFactoryConstants.SAMPLE_USER_NAME);
+		ddlRecordVersionModel.setCreateDate(dDLRecordModel.getModifiedDate());
+		ddlRecordVersionModel.setDDMStorageId(dDLRecordModel.getDDMStorageId());
+		ddlRecordVersionModel.setRecordSetId(dDLRecordModel.getRecordSetId());
+		ddlRecordVersionModel.setRecordId(dDLRecordModel.getRecordId());
+		ddlRecordVersionModel.setVersion(dDLRecordModel.getVersion());
+		ddlRecordVersionModel.setDisplayIndex(dDLRecordModel.getDisplayIndex());
+		ddlRecordVersionModel.setStatus(WorkflowConstants.STATUS_APPROVED);
+		ddlRecordVersionModel.setStatusDate(dDLRecordModel.getModifiedDate());
+
+		return ddlRecordVersionModel;
+	}
+
+	public DDMContentModel newDDMContentModel(
+		DDLRecordModel ddlRecordModel, int currentIndex) {
+
+		StringBundler sb = new StringBundler(
+			3 + BenchmarksPropsValues.MAX_DDL_CUSTOM_FIELD_COUNT * 7);
+
+		sb.append("{\"availableLanguageIds\": [\"en_US\"],");
+		sb.append("\"defaultLanguageId\": \"en_US\", \"fieldValues\": [");
+
+		for (int i = 0; i < BenchmarksPropsValues.MAX_DDL_CUSTOM_FIELD_COUNT;
+			 i++) {
+
+			sb.append("{\"instanceId\": \"");
+			sb.append(StringUtil.randomId());
+			sb.append("\", \"name\": \"");
+			sb.append(nextDDLCustomFieldName(ddlRecordModel.getGroupId(), i));
+			sb.append("\", \"value\": {\"en_US\": \"Test Record ");
+			sb.append(currentIndex);
+			sb.append("\"}},");
+		}
+
+		if (BenchmarksPropsValues.MAX_DDL_CUSTOM_FIELD_COUNT > 0) {
+			sb.setIndex(sb.index() - 1);
+		}
+
+		sb.append("]}");
+
+		return newDDMContentModel(
+			ddlRecordModel.getDDMStorageId(), ddlRecordModel.getGroupId(),
+			sb.toString());
+	}
+
+	public DDMStructureLinkModel newDDMStructureLinkModel(
+		DDLRecordSetModel ddlRecordSetModel) {
+
+		return newDDMStructureLinkModel(
+			getClassNameId(DDLRecordSet.class),
+			ddlRecordSetModel.getRecordSetId(),
+			ddlRecordSetModel.getDDMStructureId());
+	}
+
+	public DDMStructureLinkModel newDDMStructureLinkModel(
+		DLFileEntryMetadataModel dLFileEntryMetadataModel) {
+
+		return newDDMStructureLinkModel(
+			getClassNameId(DLFileEntryMetadata.class),
+			dLFileEntryMetadataModel.getFileEntryMetadataId(),
+			dLFileEntryMetadataModel.getDDMStructureId());
+	}
+
+	public DDMStructureVersionModel newDDMStructureVersionModel(
+		DDMStructureModel ddmStructureModel) {
+
+		return newDDMStructureVersionModel(
+			ddmStructureModel, _defaultDDLDDMStructureVersionId);
+	}
+
+	protected DDMStructureLinkModel newDDMStructureLinkModel(
+		long classNameId, long classPK, long structureId) {
+
+		DDMStructureLinkModel ddmStructureLinkModel =
+			new DDMStructureLinkModelImpl();
+
+		ddmStructureLinkModel.setStructureLinkId(counter.get());
+		ddmStructureLinkModel.setClassNameId(classNameId);
+		ddmStructureLinkModel.setClassPK(classPK);
+		ddmStructureLinkModel.setStructureId(structureId);
+
+		return ddmStructureLinkModel;
+	}
+
+	protected String nextDDLCustomFieldName(
+		long groupId, int customFieldIndex) {
+
+		StringBundler sb = new StringBundler(4);
+
+		sb.append("custom_field_text_");
+		sb.append(groupId);
+		sb.append("_");
+		sb.append(customFieldIndex);
+
+		return sb.toString();
+	}
+
+	private final long _defaultDDLDDMStructureVersionId;
+
+}
