+++ modules/util/portal-tools-sample-sql-builder/src/main/java/com/liferay/portal/tools/sample/sql/builder/SampleSQLBuilder.java
@@ -253,6 +253,8 @@ public class SampleSQLBuilder {
 						).put(
 							"dataFactory", new DataFactory()
 						).put(
+							"dDLDDMDataFactory", new DDLDDMDataFactory()
+						).put(
 							"dLDataFactory", new DLDataFactory()
 						).put(
 							"journalDataFactory", new JournalDataFactory()
